requirements.txt
alpaca-trade-api>=2.3.0
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=5.4.1
SQLAlchemy>=1.4.0
pymysql>=1.0.2
python-dotenv>=0.19.0